chore(overview): prefetching less aggressive
