+++ src/containers/ProjectTreeTable/ProjectTreeTable.tsx
@@ -40,11 +40,11 @@ import { ClipboardProvider } from './context/ClipboardContext'
 import useCustomColumnWidthVars from './hooks/useCustomColumnWidthVars'
 
 // Utility function imports
-import { getCellId, parseCellId } from './utils/cellUtils'
+import { getCellId } from './utils/cellUtils'
 import useLocalStorage from '@hooks/useLocalStorage'
 import { useProjectTableContext } from '@containers/ProjectTreeTable/context/ProjectTableContext'
 import useCellContextMenu from './hooks/useCellContextMenu'
-import { useLazyGetTasksByParentQuery } from '@queries/overview/getOverview'
+import usePrefetchFolderTasks from './hooks/usePrefetchFolderTasks'
 
 //These are the important styles to make sticky column pinning work!
 //Apply styles like this using your CSS strategy of choice with this kind of logic to head cells, data cells, footer cells, etc.
@@ -250,8 +250,6 @@ const FlexTable = ({
     columnPinningUpdater,
     columnOrder,
     columnOrderUpdater,
-    queryFilters,
-    projectName,
   } = useProjectTableContext()
 
   // COLUMN SIZING
@@ -353,33 +351,7 @@ const FlexTable = ({
 
   const { handleTableBodyContextMenu } = useCellContextMenu({ attribs })
 
-  const [fetchFolderTasks] = useLazyGetTasksByParentQuery()
-
-  const handlePreFetchTasks = (e: React.MouseEvent<HTMLTableSectionElement>) => {
-    const target = e.target as HTMLTableCellElement
-    const td = target.closest('td')
-    // hovering td?
-    if (!td) return
-    // first div child of td id
-    const cell = td.firstElementChild as HTMLDivElement
-    const cellId = cell?.id
-    if (!cellId) return
-    // cell colId is name
-    const { colId, rowId } = parseCellId(cellId) || {}
-    if (colId !== 'name' || !rowId) return
-    // check if the cell is a folder (classname contains folder)
-    const isFolder = cell.className.includes('folder')
-    if (!isFolder) return
-    fetchFolderTasks(
-      {
-        projectName,
-        parentIds: [rowId],
-        filter: queryFilters.filterString,
-        search: queryFilters.search,
-      },
-      true,
-    )
-  }
+  const { handlePreFetchTasks } = usePrefetchFolderTasks()
 
   return (
     <Styled.TableWrapper>
